+++ src/condition/Len.js
@@ -18,15 +18,19 @@ module.exports = function (options) {
         });
 
     return utils.customMessageWrapper(function lengthCondition(value, path) {
+        var key;
+        
         if (typeof value !== 'string' && !Array.isArray(value)) {
             throw new ConditionViolationException(value, path, messages.type, opts);
         }
         
         if (value.length > opts.max || (value.length === opts.max && !opts.maxInclusive)) {
-            throw new ConditionViolationException(value, path, messages.max, opts);
+            key = 'max' + (opts.maxInclusive ? 'Inclusive' : '');
+            throw new ConditionViolationException(value, path, messages[key], opts);
         }
         if (value.length < opts.min || (value.length === opts.min && !opts.minInclusive)) {
-            throw new ConditionViolationException(value, path, messages.min, opts);
+            key = 'min' + (opts.minInclusive ? 'Inclusive' : '');
+            throw new ConditionViolationException(value, path, messages[key], opts);
         }
     });
 };