+++ src/messages.js
@@ -1,10 +1,13 @@
 module.exports = {
     'Array': 'must be an array',
+    'Custom': 'custom condition failed',
     'Integer': 'must be an integer',
     'IsoDateTime': 'not a valid ISO 8601 datetime string',
     'Len': {
         'max': 'must not be longer than ${max}',
+        'maxInclusive': 'must not be longer than or equal to ${max}',
         'min': 'must not be shorter than ${min}',
+        'minInclusive': 'must not be shorter than or equal to ${min}',
         'type': 'must be a string or array'
     },
     'Literal': 'must be equal to ${valueStr}',
@@ -17,7 +20,9 @@ module.exports = {
     },
     'Range': {
         'max': 'must not be greater than ${max}',
-        'min': 'must not be smaller than ${min}'
+        'maxInclusive': 'must not be greater than or equal to ${max}',
+        'min': 'must not be smaller than ${min}',
+        'minInclusive': 'must not be smaller than or equal to ${min}'
     },
     'Regex': 'must match the regular expression',
     'Some': 'did not satisfy any of the given conditions',