+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 Breaking changes:
 
  - `lewd.Array`'s behaviour changed from wrapping its arguments into `lewd.some` to `lewd.all`. So `[String, /^\d+$/]` now requires all elements to be strings *and* match the regular expression (instead of needing to only match either).
+ 
+New features:
+
+ - Custom conditions can now easily inlined (see the wiki) 
+
+Improvements:
+
+ - the `range` and `len` conditions have more fine-grained error messages
 
 ## 0.1.0 (2014-05-28)